refactor(mcp): extract the tool implementations into agent_tools.py
Moves the nine _execute_* handlers and their helpers out of
MCPServerWrapper into a shared PredbatTools class, with TOOL_DEFS
projected into either MCP inputSchema or OpenAI function-calling
parameters. MCPServerWrapper now inherits the handlers, so its existing
tests exercise the moved code unchanged, and a golden copy of tools/list
guards the contract.

Groundwork for the web chat agent, which needs the same tools in the
other schema dialect without speaking MCP.
